infrastructure/pyproject.toml
[project]
name = "treelemetry-infrastructure"
version = "0.1.0"
description = "AWS CDK infrastructure for Treelemetry"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["infrastructure"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]

